+++ Source/SwiftLintFramework/Rules/Lint/InertDeferRule.swift
@@ -25,7 +25,17 @@ struct InertDeferRule: ConfigurationProviderRule, SwiftSyntaxRule {
                     print("other code")
                 }
             }
-            """)
+            """),
+            Example("""
+            func f() {
+                #if os(macOS)
+                defer { print(2) }
+                #else
+                defer { print(3) }
+                #endif
+                print(1)
+            }
+            """, excludeFromDocumentation: true)
         ],
         triggeringExamples: [
             Example("""
@@ -46,7 +56,23 @@ struct InertDeferRule: ConfigurationProviderRule, SwiftSyntaxRule {
                     // comment
                 }
             }
-            """)
+            """),
+            Example("""
+            func f(arg: Int) {
+                if arg == 1 {
+                    ↓defer { print(2) }
+                    // a comment
+                } else {
+                    ↓defer { print(3) }
+                }
+                print(1)
+                #if os(macOS)
+                ↓defer { print(4) }
+                #else
+                ↓defer { print(5) }
+                #endif
+            }
+            """, excludeFromDocumentation: true)
         ]
     )
 
@@ -58,13 +84,25 @@ struct InertDeferRule: ConfigurationProviderRule, SwiftSyntaxRule {
 private extension InertDeferRule {
     final class Visitor: ViolationsSyntaxVisitor {
         override func visitPost(_ node: DeferStmtSyntax) {
-            guard let codeBlockItem = node.parent?.as(CodeBlockItemSyntax.self),
-                  let codeBlockList = codeBlockItem.parent?.as(CodeBlockItemListSyntax.self),
-                  codeBlockList.last == codeBlockItem  else {
+            guard node.isLastStatement else {
+                return
+            }
+            if let ifConfigClause = node.parent?.parent?.parent?.as(IfConfigClauseSyntax.self),
+               ifConfigClause.parent?.parent?.isLastStatement == false {
                 return
             }
 
             violations.append(node.deferKeyword.positionAfterSkippingLeadingTrivia)
         }
+    }
+}
+
+private extension SyntaxProtocol {
+    var isLastStatement: Bool {
+        if let codeBlockItem = parent?.as(CodeBlockItemSyntax.self),
+           let codeBlockList = codeBlockItem.parent?.as(CodeBlockItemListSyntax.self) {
+            return codeBlockList.last == codeBlockItem
+        }
+        return false
     }
 }